Holoscan SDK 2.5.0 Release

+++ examples/import_gxf_components/cpp/import_gxf_components.cpp
@@ -46,7 +46,7 @@
                          __LINE__,                                                      \
                          __FILE__,                                                      \
                          cudaGetErrorString(cuda_status),                               \
-                         cuda_status);                                                  \
+                         static_cast<int>(cuda_status));                                \
     }                                                                                   \
   }
 #endif
