+++ include/holoscan/utils/cuda_macros.hpp
@@ -34,7 +34,7 @@
                     __LINE__,                                                              \
                     __FILE__,                                                              \
                     cudaGetErrorString(_holoscan_cuda_err),                                \
-                    _holoscan_cuda_err);                                                   \
+                    static_cast<int>(_holoscan_cuda_err));                                 \
     }                                                                                      \
     _holoscan_cuda_err;                                                                    \
   })
@@ -48,7 +48,7 @@
                       __LINE__,                                                              \
                       __FILE__,                                                              \
                       cudaGetErrorString(_holoscan_cuda_err),                                \
-                      _holoscan_cuda_err);                                                   \
+                      static_cast<int>(_holoscan_cuda_err));                                 \
     }                                                                                        \
     _holoscan_cuda_err;                                                                      \
   })
